Paths for views have changed in J4 - Fix Components not completely listed + Fix Frontend / Backend Problem Session scope sessionroot (#12)
* Update listcomponents.php

Bugfix for Joomla 4.x - The relevant components are not listed in select-field in plugin quantummanagermedia - paths for views have changed in J4

* Update quantummodal.js

The path could not be resolved correctly for scope sessionroot because of the missing session variable with different sessions in Frontend and Backend. This fixes the false resolveing of the path.